Security feature changes

+++ src/main/java/com/example/seproject2022/controller/ProductController.java
@@ -7,6 +7,7 @@ import com.example.seproject2022.model.dto.ProductDtoForPaginationAndGroupByCate
 import com.example.seproject2022.model.dto.ProductDtoUpdate;
 import com.example.seproject2022.model.dto.PageSettings;
 import com.example.seproject2022.service.ProductService;
+import com.example.seproject2022.service.ValidatorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -19,6 +20,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -31,37 +33,50 @@ public class ProductController {
     @Autowired
     private ProductService productService;
 
+    @Autowired
+    private ValidatorService validatorService;
+
     @GetMapping()
-    public PageDto<ProductDtoForPaginationAndGroupByCategory> getProductsWithFilters(
-        @RequestBody PageSettings pageSettings) {
+    public PageDto<ProductDtoForPaginationAndGroupByCategory> getProductsWithFilters(@RequestHeader(value = "jwt", required = false) String jwt,
+                                                                                     @RequestBody PageSettings pageSettings,
+                                                                                     HttpServletRequest request) {
+        validatorService.validateIsUser(jwt, request.getRequestURI());
         Sort productSort = pageSettings.buildSort();
         Pageable productPage = PageRequest.of(pageSettings.getPage(), pageSettings.getElementPerPage(), productSort);
         return productService.getProducts(productPage);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ProductDtoCreateProductResponse> getProductById(@PathVariable("id") Long id,
-                                                                HttpServletRequest request) {
+    public ResponseEntity<ProductDtoCreateProductResponse> getProductById(@RequestHeader(value = "jwt", required = false) String jwt,
+                                                                          @PathVariable("id") Long id,
+                                                                          HttpServletRequest request) {
+        validatorService.validateIsAdmin(jwt, request.getRequestURI());
         return new ResponseEntity<>(productService.getProductById(id, request.getRequestURI()), HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ProductDtoUpdate> updateProduct(@PathVariable("id") Long id,
+    public ResponseEntity<ProductDtoUpdate> updateProduct(@RequestHeader(value = "jwt", required = false) String jwt,
+                                                          @PathVariable("id") Long id,
                                                           @RequestBody ProductDtoUpdate productDto,
                                                           HttpServletRequest request) {
+        validatorService.validateIsAdmin(jwt, request.getRequestURI());
         return new ResponseEntity<>(productService.updateProductById(id, productDto, request.getRequestURI()),
                                     HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable("id") Long id, HttpServletRequest request) {
+    public ResponseEntity<Void> deleteProduct(@RequestHeader(value = "jwt", required = false) String jwt,
+                                              @PathVariable("id") Long id, HttpServletRequest request) {
+        validatorService.validateIsAdmin(jwt, request.getRequestURI());
         productService.deleteProductById(id, request.getRequestURI());
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping()
-    public ResponseEntity<ProductDtoCreateProductResponse> createProduct(
-        @RequestBody ProductDtoCreateInput productDto) {
+    public ResponseEntity<ProductDtoCreateProductResponse> createProduct(@RequestHeader(value = "jwt", required = false) String jwt,
+                                                                         @RequestBody ProductDtoCreateInput productDto,
+                                                                         HttpServletRequest request) {
+        validatorService.validateIsAdmin(jwt, request.getRequestURI());
         return new ResponseEntity<>(productService.saveProduct(productDto), HttpStatus.OK);
     }
 }
